+++ src/main/gen/icu/windea/pls/localisation/psi/ParadoxLocalisationVisitor.java
@@ -69,10 +69,6 @@ public class ParadoxLocalisationVisitor extends PsiElementVisitor {
     visitPsiElement(o);
   }
 
-  public void visitSequentialNumber(@NotNull ParadoxLocalisationSequentialNumber o) {
-    visitRichText(o);
-  }
-
   public void visitString(@NotNull ParadoxLocalisationString o) {
     visitRichText(o);
   }