+++ src/main/kotlin/icu/windea/pls/localisation/highlighter/ParadoxLocalisationColorSettingsPage.kt
@@ -16,7 +16,6 @@ private val attributesDescriptors = arrayOf(
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.parameter"), ParadoxLocalisationAttributesKeys.PARAMETER_KEY),
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.propertyReference"), ParadoxLocalisationAttributesKeys.PROPERTY_REFERENCE_KEY),
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.icon"), ParadoxLocalisationAttributesKeys.ICON_KEY),
-	AttributesDescriptor(PlsBundle.message("localisation.displayName.sequentialNumber"), ParadoxLocalisationAttributesKeys.SEQUENTIAL_NUMBER_KEY),
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.commandScope"), ParadoxLocalisationAttributesKeys.COMMAND_SCOPE_KEY),
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.commandField"), ParadoxLocalisationAttributesKeys.COMMAND_FIELD_KEY),
 	AttributesDescriptor(PlsBundle.message("localisation.displayName.color"), ParadoxLocalisationAttributesKeys.COLOR_KEY),