+++ src/main/kotlin/icu/windea/pls/localisation/psi/ParadoxLocalisationPsiExtensions.kt
@@ -14,25 +14,8 @@ val ParadoxLocalisationPropertyReference.propertyReferenceId: PsiElement? get() 
 
 val ParadoxLocalisationIcon.iconId: PsiElement? get() = findChild(ICON_ID)
 
-val ParadoxLocalisationSequentialNumber.sequentialNumberId: PsiElement? get() = findChild(SEQUENTIAL_NUMBER_ID)
-
 val ParadoxLocalisationColorfulText.colorId: PsiElement? get() = findChild(COLOR_ID)
 
 val ParadoxLocalisationCommandScope.commandScopeId: PsiElement get() = findRequiredChild(COMMAND_SCOPE_ID)
 
 val ParadoxLocalisationCommandField.commandFieldId: PsiElement? get() = findChild(COMMAND_FIELD_ID)
-
-fun ParadoxLocalisationLocale.changeLocaleId(id: String) {
-	val newId = ParadoxLocalisationElementFactory.createLocale(project, id).localeId
-	this.localeId.replace(newId)
-}
-
-fun ParadoxLocalisationSequentialNumber.changeSequentialNumberId(id: String) {
-	val newId = ParadoxLocalisationElementFactory.createSequentialNumber(project, id).sequentialNumberId ?: return
-	this.sequentialNumberId?.replace(newId)
-}
-
-fun ParadoxLocalisationColorfulText.changeColorId(id: String) {
-	val newId = ParadoxLocalisationElementFactory.createColorfulText(project, id).colorId ?: return
-	this.colorId?.replace(newId)
-}